+++ backend/srcs/src/game/game.module.ts
@@ -4,6 +4,7 @@ import { GameGateway } from './game.gateway';
 import { HistoricGamesService } from 'src/historic_games/historic_games.service';
 import { AuthService } from 'src/auth/auth.service';
 import { JwtService } from "@nestjs/jwt";
+import { UserService } from 'src/user/user.service';
 
 @Module({
   providers: [
@@ -11,7 +12,8 @@ import { JwtService } from "@nestjs/jwt";
     GameGateway,
     HistoricGamesService,
     AuthService,
-    JwtService
+    JwtService, 
+    UserService
   ]
 })
 export class GameModule {}