+++ backend/srcs/src/tfa/tfa.service.ts
@@ -1,7 +1,6 @@
-import { Injectable } from '@nestjs/common';
+import { HttpStatus, Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { TwoFA } from '../auth/auth.2fa'
-import { User } from "@prisma/client"
 import * as QRCode from 'qrcode';
 import * as speakeasy from 'speakeasy';
 import { ValidateDto } from "./dto";
@@ -14,7 +13,7 @@ export class TfaService {
     try {
       const secret = TwoFA.generateSecret();
 
-      await this.prisma.user.update({
+      const user = await this.prisma.user.update({
         where: {
           userId: UserId,
         },
@@ -22,13 +21,13 @@ export class TfaService {
           twofa_code: secret,
           twofa: true,
         },
-      })
+      });
 
       const otpauthUrl = speakeasy.otpauthURL({
         secret: secret,
         encoding: "base64",
-        label: "Tomartin's Game",
-        issuer: 'NombreDeTuAplicacion',
+        label: user.username,
+        issuer: 'ft_transcendence',
       });
 
       const dataURL = await QRCode.toDataURL(otpauthUrl, { errorCorrectionLevel: 'H', encoding: 'base64' });
@@ -59,12 +58,33 @@ export class TfaService {
       });
 
       if (isValid) {
-        return res.status(200).json({ message: 'Código de autenticación válido' });
+        return res.status(HttpStatus.OK).json({ message: 'Código de autenticación válido' });
       } else {
-        return res.status(401).json({ error: 'Código de autenticación inválido' });
+        return res.status(HttpStatus.UNAUTHORIZED).json({ error: 'Código de autenticación inválido' });
       }
     } catch (error) {
-      return res.status(401).json({ error: 'Código de autenticación inválido' });
+      return res.status(HttpStatus.UNAUTHORIZED).json({ error: 'Código de autenticación inválido' });
+    }
+  }
+
+  async   validTFACode(body: ValidateDto, res: any): Promise<Boolean> {
+    try {
+      const secret: { twofa_code: string } = await this.prisma.user.findUnique({
+        where: {
+          userId: body.userid,
+        },
+        select: {
+          twofa_code: true,
+        }
+      })
+      const isValid = speakeasy.totp.verify({
+        secret: secret.twofa_code,
+        encoding: 'base64',
+        token: body.token,
+      });
+      return isValid;
+    } catch (error) {
+      return res.status(HttpStatus.UNAUTHORIZED).json({ response: 'Unauthorized'});
     }
   }
 }