+++ backend/srcs/src/utils/token.constans.ts
@@ -1,5 +1,6 @@
 export class TokenConstants {
     public static readonly USER_TOKEN = 'USER_TOKEN';
     public static readonly REFRESH_TOKEN = 'REFRESH_TOKEN';
+    public static readonly TFA_TOKEN = 'TFA_TOKEN';
   }