+++ frontend/srcs/src/app/app.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { AuthService } from './services';
 
 @Component({
   selector: 'app-root',
@@ -7,5 +8,12 @@ import { Component } from '@angular/core';
 })
 export class AppComponent {
   darkMode: boolean = true;
+  constructor(
+    private readonly authService: AuthService
+  ){ }
+
+  public isLoggedIn() {
+    return this.authService.isLoggedIn();
+  }
 
 }