+++ frontend/srcs/src/app/core/api.interceptor.ts
@@ -4,11 +4,13 @@ import { Router } from '@angular/router';
 import { EMPTY } from 'rxjs';
 import { RefreshTokenManageService } from '../services/refresh-token-manager.service';
 import { UriConstants } from '../utils';
+import { environment } from 'src/environments/environment';
 
 export const ApiInterceptor = (request: HttpRequest<unknown>, next: HttpHandlerFn) => {
-	// if (request.url === URL_AUTH_SIGNIN) {
-	// 	return next(request);
-	// }
+	if (request.url === UriConstants.VALID_TFA || 
+		request.url === environment.loginUrl) {
+		return next(request);
+	}
 
 	const refreshTokenManageService = inject(RefreshTokenManageService);