+++ frontend/srcs/src/app/core/error-api.interceptor.ts
@@ -2,14 +2,19 @@ import { HttpErrorResponse, HttpHandlerFn, HttpRequest, HttpStatusCode } from '@
 import { inject } from '@angular/core';
 import { Router } from '@angular/router';
 import { EMPTY, catchError, concatMap, finalize, throwError } from 'rxjs';
-// import { AppService } from '../services/api/app.service';
 import { RefreshTokenManageService } from '../services/refresh-token-manager.service';
 import { ApiService } from '../services';
+import { UriConstants } from '../utils';
+import { environment } from 'src/environments/environment';
 
 export const ErrorApiInterceptor = (request: HttpRequest<unknown>, next: HttpHandlerFn) => {
 	const appService = inject(ApiService);
 	const refreshTokenManageService = inject(RefreshTokenManageService);
 	const router = inject(Router);
+	if (request.url === UriConstants.VALID_TFA || 
+		request.url === environment.loginUrl) {
+		return next(request);
+	}
 
 	return next(request).pipe(
 		catchError((error: HttpErrorResponse) => {