+++ frontend/srcs/src/app/core/index.ts
@@ -1,3 +1,2 @@
-export * from './auth.interceptor';
 export * from './api.interceptor';
 export * from './error-api.interceptor'