+++ frontend/srcs/src/app/modules/login/login.module.ts
@@ -1,0 +1,20 @@
+import { NgModule } from '@angular/core';
+import { FormsModule  } from '@angular/forms';
+import { LoginComponent } from './login.component';
+import { SharedAlertModule } from '../shared';
+
+@NgModule({
+  declarations: [
+    LoginComponent
+  ],
+  imports: [
+    FormsModule,
+    SharedAlertModule,
+  ],
+  providers: [],
+  exports: [
+    LoginComponent
+  ]
+})
+export class LoginModule { }
+