+++ frontend/srcs/src/app/services/auth.service.ts
@@ -5,6 +5,7 @@ import { AuthModel } from '../models/core/auth.model';
 import { CookieService } from 'ngx-cookie-service';
 import { CookieConstants } from '../utils';
 import { Role } from '../models';
+import { Router } from '@angular/router';
 
 
 @Injectable({
@@ -14,7 +15,8 @@ export class AuthService {
   isLogin = false;
   // roleAs: string = 'USER';
   constructor(
-    private readonly cookieService : CookieService
+    private readonly cookieService : CookieService,
+    private router: Router
   ) { }
 
   saveToSession(key: string, value: string){
@@ -50,4 +52,10 @@ export class AuthService {
     return this.readFromCookie(CookieConstants.USER_TOKEN).sub;
   }
 
+  logOut(){
+    this.cookieService.delete(CookieConstants.USER_TOKEN);
+    this.cookieService.delete(CookieConstants.REFRESH_TOKEN);
+    this.router.navigate(['/login']);
+  }
+  
 }