+++ frontend/srcs/src/app/utils/cookie.constants.ts
@@ -1,6 +1,7 @@
 export class CookieConstants {
   public static readonly USER_TOKEN = 'USER_TOKEN';
   public static readonly REFRESH_TOKEN = 'REFRESH_TOKEN';
+  public static readonly TFA_TOKEN = 'TFA_TOKEN';
   // public static readonly MENUS = 'MENUS';
 
 }