+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2023-11-23 - emadriga
+### Added
+- Login/logout complete workflow using tokens, 2fa & roles
+
 ## 2023-11-22 - jvacaris
 ### Added
 - Game ends well, posts the result to our API and allows clients to start another match.
@@ -13,7 +17,6 @@ and this project adheres to
 
 ### To do
 - Check if a player is already in a match before accessing the matchmaking.
-
 
 ## 2023-11-21 - emadriga
 ### Changed